More timestamp

+++ Assets/Scripts/UI/UIManager.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -44,7 +45,7 @@ public class UIManager : MonoBehaviour
 
     public static UIManager instance;
 
-    public  static int debugTimeStamp;
+    public  static long debugTimeStamp;
     [SerializeField]
     private Text debugText;
     public static void StartGame()
@@ -86,7 +87,7 @@ public class UIManager : MonoBehaviour
     private void Update()
     {
         GameTimer.instance?.OnUpdate();
-        debugText.text = ""+debugTimeStamp;
+        debugText.text = "" + debugTimeStamp;
     }
     private void OnDestroy()
     {
